+++ vortex_utils/CMakeLists.txt
@@ -5,15 +5,8 @@ set(CMAKE_CXX_STANDARD 20)
 add_compile_options(-Wall -Wextra -Wpedantic)
 
 find_package(ament_cmake REQUIRED)
-find_package(rclcpp REQUIRED)
 find_package(ament_cmake_python REQUIRED)
 find_package(Eigen3 REQUIRED)
-find_package(tf2 REQUIRED)
-find_package(tf2_ros REQUIRED)
-find_package(tf2_geometry_msgs REQUIRED)
-find_package(geometry_msgs REQUIRED)
-
-include_directories(include)
 
 add_library(${PROJECT_NAME} SHARED
   src/math.cpp)
@@ -23,12 +16,7 @@ target_include_directories(${PROJECT_NAME} PUBLIC
   $<INSTALL_INTERFACE:include>)
 
 ament_target_dependencies(${PROJECT_NAME} PUBLIC
-  rclcpp
   Eigen3
-  tf2
-  tf2_ros
-  tf2_geometry_msgs
-  geometry_msgs
 )
 
 ament_export_targets(${PROJECT_NAME} HAS_LIBRARY_TARGET)
