+++ vortex_utils/cpp_test/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 3.8)
 
 find_package(GTest REQUIRED)
+find_package(Eigen3 REQUIRED)
 include(GoogleTest)
 
 set(TEST_BINARY_NAME ${PROJECT_NAME}_test)
@@ -9,7 +10,6 @@ add_executable(
     test_math.cpp
     test_types.cpp
     test_concepts.cpp
-    test_ros_conversions.cpp
 )
 
 target_link_libraries(
@@ -23,26 +23,3 @@ target_link_libraries(
 ament_target_dependencies(${TEST_BINARY_NAME} PUBLIC Eigen3)
 
 gtest_discover_tests(${TEST_BINARY_NAME})
-
-set(TEST_BINARY_ROS_NAME ${PROJECT_NAME}_ros_test)
-
-add_executable(
-  ${TEST_BINARY_ROS_NAME}
-  test_ros_transforms.cpp
-)
-
-target_link_libraries(
-  ${TEST_BINARY_ROS_NAME}
-    ${PROJECT_NAME}
-    GTest::GTest
-)
-
-ament_target_dependencies(
-  ${TEST_BINARY_ROS_NAME}
-  rclcpp
-  tf2
-  tf2_ros
-  tf2_geometry_msgs
-)
-
-gtest_discover_tests(${TEST_BINARY_ROS_NAME})
