+++ vortex_utils/include/vortex/utils/math.hpp
