+++ vortex_utils/src/math.cpp
