+++ vortex_utils_ros/cpp_test/CMakeLists.txt
@@ -1,0 +1,23 @@
+cmake_minimum_required(VERSION 3.8)
+
+find_package(GTest REQUIRED)
+include(GoogleTest)
+
+set(TEST_BINARY_NAME ${PROJECT_NAME}_test)
+
+add_executable(${TEST_BINARY_NAME}
+  test_ros_conversions.cpp
+)
+
+target_link_libraries(${TEST_BINARY_NAME}
+  PRIVATE
+  ${PROJECT_NAME}
+  GTest::GTest
+  GTest::gtest_main
+)
+
+ament_target_dependencies(${TEST_BINARY_NAME} PUBLIC
+  rclcpp
+)
+
+gtest_discover_tests(${TEST_BINARY_NAME})
