+++ vortex_utils_ros/package.xml
@@ -1,22 +1,20 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>vortex_utils</name>
+  <name>vortex_utils_ros</name>
   <version>0.0.0</version>
-  <description>Package containing utility functions and data structures.</description>
-  <maintainer email="andeshog@stud.ntnu.no">andeshog</maintainer>
+  <description>Package containing ROS utility functions and data structures.</description>
+  <maintainer email="jorgen.fjermedal@hotmail.com">jorgenfj</maintainer>
   <license>MIT</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>ament_cmake_python</buildtool_depend>
 
   <depend>rclcpp</depend>
+  <depend>vortex_utils</depend>
   <depend>eigen</depend>
   <depend>python3-numpy</depend>
   <depend>python3-scipy</depend>
-  <depend>tf2</depend>
-  <depend>tf2_ros</depend>
-  <depend>tf2_geometry_msgs</depend>
   <depend>geometry_msgs</depend>
