+++ vortex_utils_ros/py_test/CMakeLists.txt
@@ -1,0 +1,13 @@
+cmake_minimum_required(VERSION 3.8)
+
+find_package(ament_cmake_pytest REQUIRED)
+set(_pytest_tests
+    test_utils_ros.py
+)
+foreach(_test_path ${_pytest_tests})
+    get_filename_component(_test_name ${_test_path} NAME_WE)
+    ament_add_pytest_test(${_test_name} ${_test_path}
+        APPEND_ENV PYTHONPATH=${CMAKE_CURRENT_BINARY_DIR}
+        WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
+    )
+endforeach()
