+++ vortex_utils_ros_tf/cpp_test/CMakeLists.txt
@@ -1,0 +1,25 @@
+cmake_minimum_required(VERSION 3.8)
+
+find_package(GTest REQUIRED)
+include(GoogleTest)
+
+set(TEST_BINARY_TF_NAME ${PROJECT_NAME}_test)
+
+
+add_executable(${TEST_BINARY_TF_NAME}
+  test_ros_transforms.cpp
+)
+
+target_link_libraries(${TEST_BINARY_TF_NAME}
+  ${PROJECT_NAME}
+  GTest::GTest
+)
+
+ament_target_dependencies(${TEST_BINARY_TF_NAME}
+  rclcpp
+  tf2
+  tf2_ros
+  tf2_geometry_msgs
+)
+
+gtest_discover_tests(${TEST_BINARY_TF_NAME})
